reset back to level graphics
