+++ code/tile.py
@@ -1,0 +1,11 @@
+import pygame 
+from settings import *
+import os
+
+class Tile(pygame.sprite.Sprite):
+	def __init__(self,pos,groups):
+		super().__init__(groups)
+		print(os.getcwd())
+		self.image = pygame.image.load('../graphics/test/rock.png').convert_alpha()
+		self.rect = self.image.get_rect(topleft = pos)
+		self.hitbox = self.rect.inflate(0,-10)
